// app/lib/sdk/validations/follow/follow.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Social API
 * OpenAPI spec version: v1
 */
import * as zod from 'zod';


export const PostApiFollowFollowBody = zod.object({
  "followerId": zod.string().optional(),
  "targetUserId": zod.string().optional()
})

export const PostApiFollowFollowResponse = zod.unknown()

export const PostApiFollowUnfollowBody = zod.object({
  "followerId": zod.string().optional(),
  "targetUserId": zod.string().optional()
})

export const PostApiFollowUnfollowResponse = zod.unknown()

export const PostApiFollowAcceptFollowRequestBody = zod.object({
  "followerId": zod.string().optional(),
  "targetUserId": zod.string().optional()
})

export const PostApiFollowAcceptFollowRequestResponse = zod.unknown()

export const PostApiFollowRejectFollowRequestBody = zod.object({
  "followerId": zod.string().optional(),
  "targetUserId": zod.string().optional()
})

export const PostApiFollowRejectFollowRequestResponse = zod.unknown()

export const getApiFollowFollowersQueryPageDefault = 1;
export const getApiFollowFollowersQueryLimitDefault = 20;

export const GetApiFollowFollowersQueryParams = zod.object({
  "page": zod.number().int().default(getApiFollowFollowersQueryPageDefault),
  "limit": zod.number().int().default(getApiFollowFollowersQueryLimitDefault)
})

export const GetApiFollowFollowersResponse = zod.unknown()

export const getApiFollowFollowingQueryPageDefault = 1;
export const getApiFollowFollowingQueryLimitDefault = 20;

export const GetApiFollowFollowingQueryParams = zod.object({
  "page": zod.number().int().default(getApiFollowFollowingQueryPageDefault),
  "limit": zod.number().int().default(getApiFollowFollowingQueryLimitDefault)
})

export const GetApiFollowFollowingResponse = zod.unknown()

export const getApiFollowPendingFollowRequestsQueryPageDefault = 1;
export const getApiFollowPendingFollowRequestsQueryLimitDefault = 20;

export const GetApiFollowPendingFollowRequestsQueryParams = zod.object({
  "page": zod.number().int().default(getApiFollowPendingFollowRequestsQueryPageDefault),
  "limit": zod.number().int().default(getApiFollowPendingFollowRequestsQueryLimitDefault)
})

export const GetApiFollowPendingFollowRequestsResponse = zod.unknown()